# applications/scmRest/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Warehouse
host: localhost:8080
basePath: /
#
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /shippings:
    x-swagger-router-controller: index
    post:
      description: Add a new shipping to the list
      tags:
        - Shippings
      operationId: createShipping
      parameters:
        - name: asset
          description: shipping properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Shipping"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/Shipping"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Update a shipping
      tags:
        - Shippings
      operationId: updateShipping
      parameters:
        - name: shipping
          description: shipping properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Shipping"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/Response"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      description: Returns a list of all shippings
      tags:
        - Shippings
      operationId: getAllShippings
      parameters:
        - name: assetClass
          in: query
          description: assetClass
          required: false
          type: string
          default: org.warehousenet.shipping
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Shippings"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /shippings/{Id}:    
    x-swagger-router-controller: index
    get:
      description: Returns a specific shipping
      tags:
        - Shippings
      operationId: getShippingById
      parameters:
        - in: path
          name: Id
          description: shippingId
          type: string
          required: true
        - name: assetClass
          in: query
          description: assetClass
          required: false
          type: string
          default: org.warehousenet.shipping
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Shipping"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Returns a specific shipping
      tags:
        - Shippings
      operationId: deleteShippingbyId
      parameters:
        - in: path
          name: Id
          description: shippingId
          type: string
          required: true
        - name: assetClass
          in: query
          description: assetClass
          required: false
          type: string
          default: org.warehousenet.shipping
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Shipping"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /shippings/history/{Id}:    
    x-swagger-router-controller: index
    get:
      description: Returns history of a shipping
      tags:
        - Shippings
      operationId: getShippingHistory
      parameters:
        - in: path
          name: Id
          description: shippingId
          type: string
          required: true
        - name: assetClass
          in: query
          description: assetClass
          required: false
          type: string
          default: org.warehousenet.shipping
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Shippings"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /stocks:
    x-swagger-router-controller: index
    post:
      description: Add a new stock to the list
      tags:
        - Stocks
      operationId: createStock
      parameters:
        - name: asset
          description: stock properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Stock"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/Stock"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Update a stock
      tags:
        - Stocks
      operationId: updateStock
      parameters:
        - name: stock
          description: stock properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Stock"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/Response"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      description: Returns a list of all stocks
      tags:
        - Stocks
      operationId: getAllStocks
      parameters:
        - name: assetClass
          in: query
          description: assetClass
          required: false
          type: string
          default: org.warehousenet.stock
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Stocks"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /stocks/{Id}:    
    x-swagger-router-controller: index
    get:
      description: Returns a specific stock
      tags:
        - Stocks
      operationId: getStockById
      parameters:
        - in: path
          name: Id
          description: stockId
          type: string
          required: true
        - name: assetClass
          in: query
          description: assetClass
          required: false
          type: string
          default: org.warehousenet.stock
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Stock"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Returns a specific stock
      operationId: deketeStockById
      tags:
        - Stocks
      parameters:
        - in: path
          name: Id
          description: stockId
          type: string
          required: true
        - name: assetClass
          in: query
          description: assetClass
          required: false
          type: string
          default: org.warehousenet.stock
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Stock"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /stocks/history/{Id}:    
    x-swagger-router-controller: index
    get:
      description: Returns history of a stock
      tags:
        - Stocks
      operationId: getStockHistory
      parameters:
        - in: path
          name: Id
          description: stockId
          type: string
          required: true
        - name: assetClass
          in: query
          description: assetClass
          required: false
          type: string
          default: org.warehousenet.stock
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Shippings"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /forecast:
      x-swagger-router-controller: index
      post:
        description: Add a new forecast to the list
        tags:
          - Forecast
        operationId: createForecast
        parameters:
          - name: asset
            description: forecast properties
            in: body
            required: true
            schema:
              $ref: "#/definitions/Forecast"
        responses:
          "201":
            description: Success
            schema:
              $ref: "#/definitions/Forecast"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
      put:
        description: Update a forecast
        tags:
          - Forecast
        operationId: updateForecast
        parameters:
          - name: asset
            description: forecast properties
            in: body
            required: true
            schema:
              $ref: "#/definitions/Forecast"
        responses:
          "201":
            description: Success
            schema:
              $ref: "#/definitions/Forecast"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
      get:
        description: Returns a list of all stocks
        tags:
          - Forecast
        operationId: getAllForecasts
        parameters:
          - name: assetClass
            in: query
            description: assetClass
            required: false
            type: string
            default: org.warehousenet.forecast
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/Forecasts"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /forecasts/{Id}:    
    x-swagger-router-controller: index
    get:
      description: Returns a specific forecast
      tags:
        - Forecast
      operationId: getforecastById
      parameters:
        - in: path
          name: Id
          description: forecastId
          type: string
          required: true
        - name: assetClass
          in: query
          description: assetClass
          required: false
          type: string
          default: org.warehousenet.forecast
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Forecast"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Returns a specific forecast
      operationId: deketeforecastById
      tags:
        - Forecast
      parameters:
        - in: path
          name: Id
          description: forecastId
          type: string
          required: true
        - name: assetClass
          in: query
          description: assetClass
          required: false
          type: string
          default: org.warehousenet.forecast
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Forecast"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /forecasts/history/{Id}:    
    x-swagger-router-controller: index
    get:
      description: Returns history of a forecast
      tags:
        - Forecast
      operationId: getForecastHistory
      parameters:
        - in: path
          name: Id
          description: forecastId
          type: string
          required: true
        - name: assetClass
          in: query
          description: assetClass
          required: false
          type: string
          default: org.warehousenet.forecast
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Forecasts"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /invoices:
      x-swagger-router-controller: index
      post:
        description: Add a new invoice to the list
        tags:
          - Invoices
        operationId: createInvoice
        parameters:
          - name: asset
            description: stock properties
            in: body
            required: true
            schema:
              $ref: "#/definitions/Invoice"
        responses:
          "201":
            description: Success
            schema:
              $ref: "#/definitions/Invoice"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
            
  /swagger:
    x-swagger-pipe: swagger_raw
  /signin:    
    x-swagger-router-controller: index
    get:
      description: Get user Token
      tags:
        - User
      operationId: signIn
      parameters:
        - name: userName
          in: query
          description: userName
          required: true
          type: string
          default: User1@org1.example.com
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Response"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /signout:    
    x-swagger-router-controller: index
    get:
      description: Returns a specific stock
      tags:
        - User
      operationId: signOut
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Response"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

definitions:
  Shipping:
    properties:
      class:
        type: string
        description: Unique identifier representing a shipping
        example: org.warehousenet.shipping
      shippingId:
        type: string
        description: Unique identifier representing a shipping
      materialId:
        type: string
        description: Unique identifier representing a material
      supplierId:
        type: string
        description: Unique identifier representing a supplier
      quantity:
        type: string
        description: Supply quantity of the shipping
      status:
        type: string
        description: Status of the shipping (e.g. sent)
    required:
      - class
      - shippingId
      - materialId
      - supplierId
  Shippings:
    type: array
    items:
      $ref: '#/definitions/Shipping'

  Stock:
    properties:
      class:
        type: string
        description: Unique identifier representing a shipping
        example: org.warehousenet.stock
      stockId:
        type: string
        description: Unique identifier representing a stock
      materialId:
        type: string
        description: Unique identifier representing a material
      supplierId:
        type: string
        description: Unique identifier representing a supplier
      quantity:
        type: string
        description: Current quantity of the stock
      min:
        type: string
        description: Minimal target quantity
      max:
        type: string
        description: Maximal target quantity
    required:
      - stockId
      - materialId
      - supplierId
  Stocks:
    type: array
    items:
      $ref: '#/definitions/Stock'

  Forecast:
    properties:
      class:
        type: string
        description: Class representing forecasts
        example: org.warehousenet.forecast
      forecastId:
        type: string
        description: Unique identifier representing a forecast
      materialId:
        type: string
        description: Unique identifier representing a material
      supplierId:
        type: string
        description: Unique identifier representing a supplier
      location:
        type: string
        description: Current quantity of the stock
      monthlyForecast:
        $ref: '#/definitions/monthlyForecastArray'
    required:
      - forecastId
      - materialId
      - supplierId
  Forecasts:
    type: array
    items:
      $ref: '#/definitions/Forecast'
      
  monthlyForecast:
    properties:
      month:
        type: string
        description: Unique identifier representing a shipping
      demand:
        type: string
        description: Unique identifier representing a forecast
      me:
        type: string
        description: Unique identifier representing a material
  monthlyForecastArray:
    type: array
    items:
      $ref: '#/definitions/monthlyForecast'
      
  Invoice:
      properties:
        class:
          type: string
          description: Unique identifier representing a shipping
          example: org.warehousenet.invoice
        collection:
          type: string
          description: collection to use
          example: invoiceCollection
        invoiceId:
          type: string
          description: Unique identifier representing a invoice
        shippingId:
          type: string
          description: Unique identifier representing a shipping
        supplierId:
          type: string
          description: Unique identifier representing a supplier
        data:
          type: string
          description: Current quantity of the stock
      required:
        - invoiceId
        - shippingId
        - supplierId
        - data

  Response:
    type: object
    properties:
      success:
        type: number
      description:
        type: string
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string